Renamed navigation tabs folder to Tabs
