+++ src/ui/widgets/Tabs/navigationTabs.module.css
